refactor: cosign keys moved to utils
